+++ pip_audit/_dependency_source/requirement.py
@@ -14,12 +14,7 @@ from typing import IO, Dict, Iterator, List, Set, Tuple, cast
 from packaging.requirements import Requirement
 from packaging.specifiers import SpecifierSet
 from packaging.version import Version
-from pip_requirements_parser import (
-    EditableRequirement,
-    InstallRequirement,
-    InvalidRequirementLine,
-    RequirementsFile,
-)
+from pip_requirements_parser import InstallRequirement, InvalidRequirementLine, RequirementsFile
 
 from pip_audit._dependency_source import (
     DependencyFixError,
@@ -95,13 +90,12 @@ class RequirementSource(DependencySource):
                 req_names: Set[str] = set()
                 for req in rf.requirements:
                     if req.req is None:
-                        # For editable requirements that don't have an egg fragment that lists the
-                        # the package name and version, `pip-requirements-parser` won't attach a
+                        # For URL requirements that don't have an egg fragment that lists the
+                        # package name and version, `pip-requirements-parser` won't attach a
                         # `Requirement` object to the `InstallRequirement`.
                         #
                         # In this case, we can't audit the dependency so we should signal to the
                         # caller that we're skipping it.
-                        assert isinstance(req, EditableRequirement)
                         yield SkippedDependency(
                             name=req.requirement_line.line,
                             skip_reason="could not deduce package/specifier pair from requirement, "